pyproject.toml
[project]
name = "pt2rknn"
version = "0.1.0"
description = "HoST Model Conversion for sim2real deployment"
requires-python = ">=3.8,<3.13"
dependencies = [
    "torch==2.0.0",
    "onnx>=1.12.0",
    "onnxruntime>=1.16.0,<1.18.0",
    "numpy<2.0.0",
    "rknn-toolkit2>=2.3.2",
    "setuptools",  # for pkg_resources
]

[tool.uv]
# Use local stub wheel for onnxoptimizer only on aarch64 (no PyPI wheel)
find-links = ["./dist"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
onnxoptimizer = [
    { path = "./dist/onnxoptimizer-0.3.8-py3-none-any.whl", marker = "platform_machine == 'aarch64'" },
]
